requirements.txt
# Core Framework
streamlit>=1.28.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# Visualization
plotly>=5.17.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Export & Reporting
openpyxl>=3.1.0
python-docx>=0.8.11
reportlab>=4.0.0
pytest>=7.4.0

# API & Networking
requests>=2.31.0
urllib3>=2.0.0

# Security & Authentication
cryptography>=41.0.0
PyJWT>=2.8.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Monitoring & Logging
python-dotenv>=1.0.0
loguru>=0.7.0

# Performance
cachetools>=5.3.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Development
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0
mypy>=1.6.0
